+++ console/src/process/form/actions.js
@@ -19,10 +19,11 @@ export const loadData = (instanceId: ConcordId, formInstanceId: ConcordId) => ({
     formInstanceId
 });
 
-export const submit = (instanceId: ConcordId, formInstanceId: ConcordId, data: mixed, wizard: boolean) => ({
+export const submit = (instanceId: ConcordId, formInstanceId: ConcordId, data: mixed, wizard: boolean, yieldFlow: boolean) => ({
     type: types.PROCESS_FORM_SUBMIT_REQUEST,
     instanceId,
     formInstanceId,
     data,
-    wizard
+    wizard,
+    yieldFlow
 });